Move all subfactors to separate module
