+++ src/__init__.py
@@ -1,0 +1,6 @@
+from cfactor.cfactor import aggregate_slr_to_c_factor, calculate_slr
+
+__all__ = [
+    "calculate_slr",
+    "aggregate_slr_to_c_factor",
+]
